fix-dto , fix-provider: auth dto moved and password ahsher  seperated
