--- backend/src/app/dtos/auth/facebook-auth.dto.ts
@@ -1,7 +1,0 @@
-export interface FacebookAuthDto {
-  accessToken: string;
-  userId: string;
-  name: string;
-  email?: string;
-  picture?: string;
-}
