--- backend/src/app/dtos/auth/forgot-password.dto.ts
@@ -1,3 +1,0 @@
-export interface ForgotPasswordDto {
-  email: string;
-}
