--- backend/src/app/dtos/auth/googel-auth.dto.ts
@@ -1,3 +1,0 @@
-export interface GoogleAuthDto {
-  accessToken: string;
-}
