--- backend/src/app/dtos/auth/index.dto.ts
@@ -1,8 +1,0 @@
-export * from "./facebook-auth.dto";
-export * from "./forgot-password.dto";
-export * from "./googel-auth.dto";
-export * from "./login.dto";
-export * from "./register.dto";
-export * from "./resend-otp.dto";
-export * from "./reset-password.dto";
-export * from "./verify-otp.dto";
