--- backend/src/app/dtos/auth/login.dto.ts
@@ -1,7 +1,0 @@
-import { AuthProvider } from "../../../domain/enum/auth-provider.enum";
-
-export interface LoginDto {
-  email: string;
-  password?: string;
-  authProvider?: AuthProvider;
-}
