--- backend/src/app/dtos/auth/register.dto.ts
@@ -1,8 +1,0 @@
-import { Role } from "../../../domain/enum/role.enum";
-
-export interface RegisterDto {
-  name: string;
-  email: string;
-  password?: string;
-  role: Role;
-}
