--- backend/src/app/dtos/auth/resend-otp.dto.ts
@@ -1,3 +1,0 @@
-export interface ResendOtpDto {
-  email: string;
-}
