--- backend/src/app/dtos/auth/reset-password.dto.ts
@@ -1,4 +1,0 @@
-export interface ResetPasswordDto {
-  resetToken: string;
-  newPassword: string;
-}
