--- backend/src/app/dtos/auth/verify-otp.dto.ts
@@ -1,5 +1,0 @@
-export interface VerifyOtpDto {
-  email: string;
-  otp: string;
-  type:string
-}
