+++ backend/src/app/providers/IPasswordHasher.ts
@@ -1,0 +1,4 @@
+export interface IPasswordHasher {
+  hash(password: string): Promise<string>;
+  compare(password: string, hashed: string): Promise<boolean>;
+}
