+++ backend/src/app/usecases/auth/implementations/facebook-auth.usecase.ts
@@ -1,11 +1,11 @@
 import { User } from "../../../../domain/entities/user.entity";
 import { IAuthRepository } from "../../../repositories/auth.repository";
 import { IFacebookAuthUseCase } from "../interfaces/facebook-auth.usecase.interface";
-import { FacebookAuthDto } from "../../../dtos/auth/facebook-auth.dto";
 import { AuthProvider } from "../../../../domain/enum/auth-provider.enum";
 import { Role } from "../../../../domain/enum/role.enum";
 import { HttpError } from "../../../../presentation/http/errors/http-error";
 import { IUpdateUserRequestDTO } from "../../../dtos/user.dto";
+import { FacebookAuthDto } from "../../../dtos/auth.dto";
 
 export class FacebookAuthUseCase implements IFacebookAuthUseCase {
   constructor(private authRepository: IAuthRepository) {}
