+++ backend/src/app/usecases/auth/implementations/forgot-passowrd.usecase.ts
@@ -1,8 +1,8 @@
 import { IAuthRepository } from "../../../repositories/auth.repository";
 import { OtpProvider } from "../../../../infra/providers/otp/otp.provider";
-import { ForgotPasswordDto } from "../../../dtos/auth/forgot-password.dto";
 import { IForgotPasswordUseCase } from "../interfaces/forgot-passowrd.usecase.interface";
 import { HttpError } from "../../../../presentation/http/errors/http-error";
+import { ForgotPasswordDto } from "../../../dtos/auth.dto";
 
 export class ForgotPasswordUseCase implements IForgotPasswordUseCase {
   constructor(
