+++ backend/src/app/usecases/auth/implementations/login.usecase.ts
@@ -3,9 +3,9 @@ import { IAuthRepository } from "../../../repositories/auth.repository";
 import { HttpError } from "../../../../presentation/http/errors/http-error";
 import * as bcrypt from "bcrypt";
 import { AuthProvider } from "../../../../domain/enum/auth-provider.enum";
-import { LoginDto } from "../../../dtos/auth/login.dto";
 import { ILoginUseCase } from "../interfaces/login.usecase.interface";
 import { ICartRepository } from "../../../repositories/cart.repository";
+import { LoginDto } from "../../../dtos/auth.dto";
 
 export class LoginUseCase implements ILoginUseCase {
   constructor(
