+++ backend/src/app/usecases/auth/implementations/reset-password.usecase.ts
@@ -2,8 +2,8 @@ import { IAuthRepository } from "../../../repositories/auth.repository";
 import { HttpError } from "../../../../presentation/http/errors/http-error";
 import * as bcrypt from "bcrypt";
 import { IResetPasswordUseCase } from "../interfaces/reset-password.usecase.interface";
-import { ResetPasswordDto } from "../../../dtos/auth/reset-password.dto";
 import { JwtProvider } from "../../../../infra/providers/auth/jwt.provider";
+import { ResetPasswordDto } from "../../../dtos/auth.dto";
 
 export class ResetPasswordUseCase implements IResetPasswordUseCase {
   constructor(private authRepository: IAuthRepository) {}
