+++ backend/src/di/shared.dependencies.ts
@@ -45,6 +45,8 @@ import { S3Service } from "../infra/providers/s3/s3.service";
 import { EmailProviderImpl } from "../infra/providers/email/email.provider";
 import { GetEnrollmentStatsUseCase } from "../app/usecases/enrollment/implementations/get-enrollment-stats.usecase";
 import { IUserRepository } from "../app/repositories/user.repository";
+import { IPasswordHasher } from "../app/providers/IPasswordHasher";
+import { BcryptPasswordHasher } from "../infra/providers/password-hasher.ts";
 
 export interface SharedDependencies {
   prisma: typeof prismaClient;
@@ -75,6 +77,7 @@ export interface SharedDependencies {
   s3Service: S3Service;
   emailProvider: EmailProviderImpl;
   getEnrollmentStatsUseCase: GetEnrollmentStatsUseCase;
+  passwordHasher: IPasswordHasher
 }
 
 export function createSharedDependencies(): SharedDependencies {
@@ -97,6 +100,7 @@ export function createSharedDependencies(): SharedDependencies {
   const walletRepository = new WalletRepository(prismaClient);
   const paymentGateway = new StripePaymentGateway();
   const webhookGateway = new StripeWebhookGateway();
+  const passwordHasher = new BcryptPasswordHasher();
 
   // Create a temporary mock revenue distribution service
   // This will be replaced with the real one in app.dependencies.ts
@@ -159,5 +163,6 @@ export function createSharedDependencies(): SharedDependencies {
     s3Service,
     emailProvider,
     getEnrollmentStatsUseCase,
+    passwordHasher,
   };
 }
