+++ backend/src/infra/providers/password-hasher.ts.ts
@@ -1,0 +1,13 @@
+import { IPasswordHasher } from "../../app/providers/IPasswordHasher";
+import * as bcrypt from 'bcrypt';
+
+
+export class BcryptPasswordHasher implements IPasswordHasher {
+  async hash(password: string): Promise<string> {
+    return bcrypt.hash(password, 10);
+  }
+
+  async compare(password: string, hashed: string): Promise<boolean> {
+    return bcrypt.compare(password, hashed);
+  }
+}
